# chore: Add caching and micro optimizations in Serializer.