--- a/src/Serializer.php
+++ b/src/Serializer.php
@@ -47,6 +47,11 @@ class Serializer
     const MAP_VALUE_FIELD_NAME = 'value';
 
     private static $phpArraySerializer;
+    // Caches for different helper functions
+    private static array $getterMap = [];
+    private static array $setterMap = [];
+    private static array $snakeCaseMap = [];
+    private static array $camelCaseMap = [];
 
     private static $metadataKnownTypes = [
         'google.rpc.retryinfo-bin' => \Google\Rpc\RetryInfo::class,
@@ -294,9 +299,10 @@ class Serializer
     {
         $data = [];
 
-        $fieldCount = $messageType->getFieldCount();
-        for ($i = 0; $i < $fieldCount; $i++) {
-            $field = $messageType->getField($i);
+        // Call the getDescriptorMaps outside of the loop to save processing.
+        // Use the same set of fields to loop over, instead of using field count.
+        list($fields, $fieldsToOneof) = $this->getDescriptorMaps($messageType);
+        foreach ($fields as $field) {
             $key = $field->getName();
             $getter = $this->getGetter($key);
             $v = $message->$getter();
@@ -306,7 +312,6 @@ class Serializer
             }
 
             // Check and skip unset fields inside oneofs
-            list($_, $fieldsToOneof) = $this->getDescriptorMaps($messageType);
             if (isset($fieldsToOneof[$key])) {
                 $oneofName = $fieldsToOneof[$key];
                 $oneofGetter =  $this->getGetter($oneofName);
@@ -433,7 +438,10 @@ class Serializer
      */
     public static function getGetter(string $name)
     {
-        return 'get' . ucfirst(self::toCamelCase($name));
+        if (! isset(self::$getterMap[$name])) {
+            self::$getterMap[$name] = 'get' . ucfirst(self::toCamelCase($name));
+        }
+        return self::$getterMap[$name];
     }
 
     /**
@@ -442,7 +450,10 @@ class Serializer
      */
     public static function getSetter(string $name)
     {
-        return 'set' . ucfirst(self::toCamelCase($name));
+        if (! isset(self::$setterMap[$name])) {
+            self::$setterMap[$name] = 'set' . ucfirst(self::toCamelCase($name));
+        }
+        return self::$setterMap[$name];
     }
 
     /**
@@ -453,7 +464,10 @@ class Serializer
      */
     public static function toSnakeCase(string $key)
     {
-        return strtolower(preg_replace(['/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'], '$1_$2', $key));
+        if (! isset(self::$snakeCaseMap[$key])) {
+            self::$snakeCaseMap[$key] = strtolower(preg_replace(['/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'], '$1_$2', $key));
+        }
+        return self::$snakeCaseMap[$key];
     }
 
     /**
@@ -464,7 +478,10 @@ class Serializer
      */
     public static function toCamelCase(string $key)
     {
-        return lcfirst(str_replace(' ', '', ucwords(str_replace('_', ' ', $key))));
+        if (! isset(self::$camelCaseMap[$key])) {
+            self::$camelCaseMap[$key] = lcfirst(str_replace(' ', '', ucwords(str_replace('_', ' ', $key))));
+        }
+        return self::$camelCaseMap[$key];
     }
 
     private static function hasBinaryHeaderSuffix(string $key)