--- a/src/Serializer.php
+++ b/src/Serializer.php
@@ -465,7 +465,9 @@ class Serializer
     public static function toSnakeCase(string $key)
     {
         if (!isset(self::$snakeCaseMap[$key])) {
-            self::$snakeCaseMap[$key] = strtolower(preg_replace(['/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'], '$1_$2', $key));
+            self::$snakeCaseMap[$key] = strtolower(
+                preg_replace(['/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'], '$1_$2', $key)
+            );
         }
         return self::$snakeCaseMap[$key];
     }