The 2 major changes here are:

### 1. Add some caching for `getGetter`, `getSetter`, `toSnakeCase` and `toCamelCase` methods.
#### Intuition:
For a simple script that calls `Select * From <table> LIMIT 100` where the table has 23 columns, the `getGetter` and `toCamelCase` were being called close to 500 times. Even though the operation is simple, but there is string manipulation and depends on the names that are used frequently. So, I simply added some caching for these methods.
### 2. Inside the `encodeMessageImpl` method, bring the call to `getDescriptorMaps` outside the loop.
#### Intuition:
The `getDescriptorMaps` depends on the `$messageType` which stays same inside the `for loop`. Moreover, the first argument returned, is actually a list of fields. So, instead of depending on the `$fieldsCount`, we can use this list to loop over the `$fields`.

### Effect
I ran a simple script that runs the same query as above 5 times, both with and without the protobuf extension enabled and I could see a good improvement for both the cases.

#### Protobuf extension not enabled

**Current(w/o this fix):**
Running query select * from Accounts LIMIT 100
Time taken: 1.73 s

Running query select * from Accounts LIMIT 100
Time taken: 1.51 s

Running query select * from Accounts LIMIT 100
Time taken: 1.63 s

Running query select * from Accounts LIMIT 100
Time taken: 1.51 s

Running query select * from Accounts LIMIT 100
Time taken: 1.49 s

Completed in 7.86 s

**With the fix:**
Running query select * from Accounts LIMIT 100
Time taken: 1.34 s

Running query select * from Accounts LIMIT 100
Time taken: 1.19 s

Running query select * from Accounts LIMIT 100
Time taken: 1.19 s

Running query select * from Accounts LIMIT 100
Time taken: 1.22 s

Running query select * from Accounts LIMIT 100
Time taken: 1.19 s

Completed in 6.14 s

#### Protobuf extension enabled
**Current(w/o this fix):**
Running query select * from Accounts LIMIT 100
Time taken: 575.849987 ms

Running query select * from Accounts LIMIT 100
Time taken: 521.261092 ms

Running query select * from Accounts LIMIT 100
Time taken: 449.397046 ms

Running query select * from Accounts LIMIT 100
Time taken: 449.931335 ms

Running query select * from Accounts LIMIT 100
Time taken: 461.896432 ms

Completed in 2.49 s

**With the fix:**

Running query select * from Accounts LIMIT 100
Time taken: 425.538723 ms

Running query select * from Accounts LIMIT 100
Time taken: 308.273615 ms

Running query select * from Accounts LIMIT 100
Time taken: 322.510244 ms

Running query select * from Accounts LIMIT 100
Time taken: 304.450003 ms

Running query select * from Accounts LIMIT 100
Time taken: 517.975938 ms

Completed in 1.88 s

Clearly, we see a good improvement in both the cases.